### 2020-06-12 #2330 @labarba
Is this JOSS submission to accompany a paper to be published with AAS journals? https://joss.readthedocs.io/en/latest/editing.html#handling-of-papers-published-together-with-aas-publishing
