### 2020-06-12 #2330 @KathrynJones1
Yes we are going to submit a paper to AJ as well
