### 2020-06-17 #2330 @arfon
:wave: @KathrynJones1, thanks for your submission to JOSS. Before we find an editor to handle this submission I wanted to ask a few questions.

This looks to be an evolution of the [`batman`](https://github.com/lkreidberg/batman) project, with the changes specific to `catwoman` here: https://github.com/lkreidberg/batman/compare/master...KathrynJones1:master . Could you say a little about the longer-term plans for this package? Is `batman` essentially deprecated now (I note there were a couple of small changes to `batman` recently but development mostly seems to have ceased).

The JOSS review process is going to inevitably include reviews a components that were developed originally for `batman` so if there's any long-term plan to merge these two codebases then reviewing `carwoman` in this way might not be optimal.

Relatedly, given the history of the codebase, it would seem appropriate to include the major contributors to `batman` as authors. Is there a reason not to include them?
