### 2020-06-18 #2330 @nespinoza
Hi @arfon, I can chime in here as we were worried by your last point even before starting this project.

To answer your first question: no, `batman` is not deprecated. People still use it, and Laura Kreidberg (the lead developer) answers to questions and comments about the code to this day. It is heavily used in the exoplanet community in fact, and it will keep being this way for a while I believe. If anything, the lack of commits in the last months/years in `batman` speaks of the good coding behind it, as no major bugs have been discovered in the code for a large time-span!

As for your second question: within `catwoman`, we wanted to use some functions in the `batman` C code but these were not easily importable. This is why we decided to write the `catwoman` algorithm by modifying the core `batman` C code in order to (a) not re-write code that is already out there, (b) give proper credit to Laura for those functions (as it is written in each of the scripts in `catwoman`) and (c) use the same code-structure and python class management that `batman` uses, so it was easier for users to use `catwoman` (as most folks in the exoplanet community are familiar with `batman`).

Before moving forward with that plan, we had a meeting with Laura in order to share it with her and to see if she was uneasy with anything we had in mind. We all agreed that `catwoman` had to be an independent endeavor to that of `batman`, because they attack different geometrical problems, despite the fact that we were going to build this taking `batman` as a base, and also because being `catwoman` a new code only tested and validated by us, it made sense that we took care of any issues that could come our way in terms of bugs, versioning, etc. We also agreed that we were going to reference `batman` both in the paper and in the actual scripts (as we do! --- the latter is actually imposed by the `batman` LICENSE, and we made sure to follow that at the finest detail). We briefly discussed authorship and decided that it would make sense to add her only if she had a direct involvement in the development of `catwoman`, which was not the case.

Given the above, I think the author list as it is right now is OK. If this makes JOSS, or you @arfon, or @lkreidberg herself uneasy at this stage, we would of course be willing to hear recommendations on this side of things. We strive for properly acknowledging the effort put on `batman`, which is on what `catwoman` was build upon.
