### 2020-06-18 #2330 @lkreidberg
Thanks for checking on this! It's totally fine with me to proceed with the JOSS paper as written. I didn't contribute at all to the new additions in `catwoman`, so I don't feel any need to be included as a coauthor - an acknowledgement would be fine.

Congrats @KathrynJones1 and @nespinoza for getting this written up!
