### 2020-06-19 #2330 @arfon
Thanks for the explanation @nespinoza and for checking in here @nespinoza, this is helpful.

I still have some concerns about the approach being taken here, specifically a general question about what this fork means for the long term of both of these packages. Our of our [submission criteria](https://joss.readthedocs.io/en/latest/submitting.html) is designed to guard against one-off modifications of projects:

> _...and designed for maintainable extension (not one-off modifications)._

but I'm not sure this applies quite here as the modification is actually describing a different transit geometry problem (and I understand why this means it can't easily be part of `batman`).

I've invited @xuanxu to edit this and I think it would be useful for him to give his opinion before deciding for sure whether we should review this.

Thanks for your patience all!
