### 2020-06-19 #2330 @nespinoza
Hi @arfon --- indeed, it is not a one-off modification. The fork in our minds just meant making sure to always let people know that we used `batman` as a base to build `catwoman`. The alternative would have been to just download the `batman` repo, do the modifications, and push that to the `catwoman` repo, which didn't seem right to us, but perhaps would have created less confusion.

In essence, right now we think of `catwoman` as a spin-off of `batman`, in that they are now two different libraries supported by independent groups. We are happy to receive any feedback on how to make all this clearer and concern-free :-).
