### 2020-06-19 #2330 @xuanxu
I see `catwoman` as a different software than `batman`, independently maintained and with no future merge plans, so I think it meets the JOSS submission criteria.

I'd also add it would be nice if the problems that made `catwoman` a fork of `batman` are addressed: so if someone wants to make use of some `catwoman` functions, it would be great if they are easily importable, using Python modules or similar to make something like this work on other people's code:
```python
import catwoman.eclipse
```
