### 2020-06-23 #2330 @xuanxu
@eschnett would you be willing to review this paper? I think your knowledge of astrophysics and C/Python makes you a great fit for this submission.
