### 2020-06-23 #2330 @eschnett
@xuanxu This paper is outside my expertise – I know next to nothing about exoplanet discovery and the methods and algorithms used there.
