### 2020-06-23 #2330 @jessie-dotson
Sure. I’d be happy to.

________________________________
On: 23 June 2020 05:56, "Juanjo Bazán" <notifications@github.com<mailto:notifications@github.com>> wrote:

@jessie-dotson<https://urldefense.proofpoint.com/v2/url?u=https-3A__github.com_jessie-2Ddotson&d=DwMCaQ&c=ApwzowJNAKKw3xye91w7BE1XMRKi2LN9kiMk5Csz9Zk&r=EsjtQGK5vNxVnulnNG39CtrT0PFIQKgMF4MiHg-NSFI&m=_73QTl0xvpO5aQe21GhuZn8-6JZJP2EHCg-VLJ8UOE4&s=dq-Wggi27WiSbYo9-RxUL4teSClncMhARAy5CzizHew&e=> would you be willing to review this paper? I think your knowledge on exoplanets and Python makes you a great fit for this submission.
