### 2020-06-23 #2382 @whedon
[REVIEW]: catwoman: A transit modelling Python package for asymmetric light curves
**Submitting author:** @KathrynJones1 (<a href="http://orcid.org/0000-0002-2316-6850">Kathryn Jones</a>)
**Repository:** <a href="https://github.com/KathrynJones1/catwoman" target ="_blank">https://github.com/KathrynJones1/catwoman</a>
**Version:** v1.0.12
