### 2020-06-25 #2382 @xuanxu
@jessie-dotson: you should have now an invitation [here](https://github.com/openjournals/joss-reviews/invitations) to collaborate in the repo, you have to accept it so you can edit your checklist.
