### 2020-06-25 KathrynJones1/catwoman#2 @barentsen
`setup.py` requires numpy
(Posted in relation to openjournals/joss-reviews#2382.)

## Issue

It is not currently possible to `pip install catwoman` in a fresh environment, because `setup.py` requires `numpy` to be installed:

```bash
$ python -m venv tmpenv
$ source tmpenv/bin/activate
$ pip install catwoman
Collecting catwoman
  Using cached https://files.pythonhosted.org/packages/b8/79/c463b58dd82d3360ce820ff263956f8804113a213b36ce7d38101275be56/catwoman-1.0.4.tar.gz
     command: /home/gb/tmp/tmpenv/bin/python3 -c 'import sys, setuptools, tokenize; sys.argv[0] = '"'"'/tmp/pip-install-unsajkma/catwoman/setup.py'"'"'; __file__='"'"'/tmp/pip-install-unsajkma/catwoman/setup.py'"'"';f=getattr(tokenize, '"'"'open'"'"', open)(__file__);code=f.read().replace('"'"'\r\n'"'"', '"'"'\n'"'"');f.close();exec(compile(code, __file__, '"'"'exec'"'"'))' egg_info --egg-base pip-egg-info
         cwd: /tmp/pip-install-unsajkma/catwoman/
    Complete output (5 lines):
    Traceback (most recent call last):
      File "<string>", line 1, in <module>
      File "/tmp/pip-install-unsajkma/catwoman/setup.py", line 5, in <module>
        import numpy as np
    ModuleNotFoundError: No module named 'numpy'
ERROR: Command errored out with exit status 1: python setup.py egg_info Check the logs for full command output.
WARNING: You are using pip version 19.2.3, however version 20.1.1 is available.
You should consider upgrading via the 'pip install --upgrade pip' command.
```

## Solution

A few different solutions exist (cf. stackoverflow).  I believe the easiest solution will be to create a `pyproject.toml` file in the root directory of the project, which is used by modern versions of `pip` to install reqiurements needed before running `setup.py`.  The file should contain something like this:
```
[build-system]
requires = ["setuptools", "wheel", "numpy"]
````
