### 2020-06-26 KathrynJones1/catwoman#1 @KathrynJones1
Hello! Thank you for this issue, I have edited the README document as you have recommended, adding a brief description of catwoman, a link to the documentation and an example of how to install. I considered copying the 'quickstart' example from the documentation but decided against it since it would be just a repeat of the Quickstart tab in the docs - however if you think this is required then let me know and I can add it in.
I have also added the long_description line to the setup.py file so it shows up on the pypi page, thank you for that recommendation!
