### 2020-06-26 KathrynJones1/catwoman#2 @KathrynJones1
Thank you for this issue!  I have had a look at the `setup.py` file and tried adding
`setup_requires=['wheel', 'numpy']` and it appeared to install when I created a new environment:

```
$ python -m venv env
$ source env/bin/activate
$ pip install catwoman
Collecting catwoman
  Downloading https://files.pythonhosted.org/packages/68/93/b0066fe24be50d80c1b33e1a064e1ff5839da703f552cc5af4cfe68cd3ac/catwoman-1.0.5-cp37-cp37m-macosx_10_7_x86_64.whl (76kB)
    100% |████████████████████████████████| 81kB 1.1MB/s
Collecting numpy (from catwoman)
  Using cached https://files.pythonhosted.org/packages/5c/f5/0e5e57fa7683cf0e5036320f4676cd7e3dbd9ab8c17ada541c2bb5ebed5e/numpy-1.19.0-cp37-cp37m-macosx_10_9_x86_64.whl
Installing collected packages: numpy, catwoman
Successfully installed catwoman-1.0.5 numpy-1.19.0
```
Let me know if this works now for you now? If not I will try the solution you suggested.
