### 2020-06-26 KathrynJones1/catwoman#2 @barentsen
Mmmm, this still doesn't work for me:

```
$ python -m venv tmpenv
$ source tmpenv/bin/activate
$ pip install catwoman
Collecting catwoman
  Downloading https://files.pythonhosted.org/packages/a4/af/e026ab937ff6baaa5b6e44b5d5a33fbbd192814deace3812a5c7897dfaa3/catwoman-1.0.5.tar.gz
     command: /home/gb/tmp/tmpenv/bin/python -c 'import sys, setuptools, tokenize;
 sys.argv[0] = '"'"'/tmp/pip-install-np_cw02a/catwoman/setup.py'"'"';
 __file__='"'"'/tmp/pip-install-np_cw02a/catwoman/setup.py'"'"';f=getattr(tokenize, '"'"'open'"'"', open)(__file__);
code=f.read().replace('"'"'\r\n'"'"', '"'"'\n'"'"');f.close();exec(compile(code, __file__, '"'"'exec'"'"'))'
egg_info --egg-base pip-egg-info
         cwd: /tmp/pip-install-np_cw02a/catwoman/
    Complete output (5 lines):
    Traceback (most recent call last):
      File "<string>", line 1, in <module>
      File "/tmp/pip-install-np_cw02a/catwoman/setup.py", line 5, in <module>
        import numpy as np
    ModuleNotFoundError: No module named 'numpy'
ERROR: Command errored out with exit status 1: python setup.py egg_info Check the logs for full command output.
```

The reason we see different behavior is that my version of `pip` attempts to install `catwoman` from source because I'm using Linux.  In your case `pip` installs the wheel file (cf. line that says `Downloading ... catwoman-1.0.5-cp37-cp37m-macosx_10_7_x86_64.whl`).  Wheel files are great because they allow `pip` to skip `setup.py` entirely and install the package by copying pre-compiled files.  You will have created those wheel files when you typed `python setup.py bdist_wheel`.  These pre-compiled wheel files are platform-specific however, and so pip will only use them for OSX users, whereas I'm using Linux.

### Solution

The reason that adding `numpy` to `setup_requires` doesn't work when installing from source is that `setup.py` contains `import numpy` at the top, so the script fails before it can inform pip about the contents of `setup_requires`.  Python introduced the `pyproject.toml` file mechanism exactly to solve this type of problem, so adding it will likely be the best solution!

(An alternative solution would be to pre-compile wheels for *every* platform, but that is way more complicated because it requires access to multi-platform build servers!)
