### 2020-06-26 KathrynJones1/catwoman#3 @barentsen
`CHANGELOG.rst` and `INSTALL` still refer to `batman`
Opened in relation to openjournals/joss-reviews#2382.

I noticed that `CHANGELOG.rst` and `INSTALL` in the root directory still belong to `batman`.  This will be confusing for users.

I recommend removing or updating these files.

Note:
* A `CHANGELOG` file is definitely very useful to have.  I recommend emptying the current one and starting a new one from scratch to log any future changes to `catwoman`.
* The `INSTALL` instructions for building the package from source could potentially be included in the online docs instead of the `INSTALL` file.
