### 2020-06-27 KathrynJones1/catwoman#2 @KathrynJones1
Ah okay I understand now, I have added the `pyproject.toml` file to the root directory and updated the pypi version - let me know if that works now!
