### 2020-06-27 KathrynJones1/catwoman#2 @barentsen
I confirm installation in a new environment works now! Awesome :+1:

```bash
$ pip install catwoman
Collecting catwoman
  Downloading https://files.pythonhosted.org/packages/95/b7/a9be27b99d2f64554a4768c1d3e2790cd4e82ddcc26e31b87df4b9e7f6b1/catwoman-1.0.6.tar.gz
  Installing build dependencies ... done
    Preparing wheel metadata ... done
Collecting numpy (from catwoman)
  Using cached https://files.pythonhosted.org/packages/57/55/5366b0afed57facaf887b05f0e98c887a9e27021e3f4e78a26a6d535e538/numpy-1.19.0-cp38-cp38-manylinux2010_x86_64.whl
Building wheels for collected packages: catwoman
  Building wheel for catwoman (PEP 517) ... done
  Created wheel for catwoman: filename=catwoman-1.0.6-cp38-cp38-linux_x86_64.whl size=111306 sha256=8c3d4e21e1ccbf3cd3466db7a6e75c859dfa4216eddc9fbfb70c1eff2fb5d405
  Stored in directory: /home/gb/.cache/pip/wheels/ff/a1/e6/50ee52f47ce80bb42fff94bb6c60d06fd8b332144e599c9000
Successfully built catwoman
Installing collected packages: numpy, catwoman
Successfully installed catwoman-1.0.6 numpy-1.19.0
```

However, importing the package doesn't work yet because the `batman` dependency is missing:

```bash
$ python -c "import catwoman"
Traceback (most recent call last):
  File "<string>", line 1, in <module>
  File "/home/gb/tmp/tmpenv/lib/python3.8/site-packages/catwoman/__init__.py", line 8, in <module>
    from .plots import *
  File "/home/gb/tmp/tmpenv/lib/python3.8/site-packages/catwoman/plots.py", line 27, in <module>
    import batman
ModuleNotFoundError: No module named 'batman'
```

It should be sufficient to add `batwoman` to `install_requires` in `setup.py` to resolve this.
