### 2020-06-30 KathrynJones1/catwoman#3 @KathrynJones1
Thank you for your recommendation! I've updated the `CHANGELOG` so that it now refers to recent `catwoman` changes. I've also removed the `INSTALL` file and added a section in the online docs to refer to this.
