### 2020-06-30 KathrynJones1/catwoman#2 @KathrynJones1
Okay, I've had a look and I've removed the `batman` dependency as it isn't necessary for the code - so that should work now? (newest version is now 1.0.7)
