### 2020-06-30 KathrynJones1/catwoman#2 @barentsen
Ok!  Installation via pip seems to work now! :+1:

```bash
$ python -m venv tmpenv
$ source tmpenv/bin/activate
$ pip install catwoman
Collecting catwoman
  Downloading https://files.pythonhosted.org/packages/bd/e0/1ab42115b03728e734effa0fa5d47b9d3854e5b7172b70a564a995cb8546/catwoman-1.0.7.tar.gz
  Installing build dependencies ... done
    Preparing wheel metadata ... done
Collecting numpy (from catwoman)
  Using cached https://files.pythonhosted.org/packages/57/55/5366b0afed57facaf887b05f0e98c887a9e27021e3f4e78a26a6d535e538/numpy-1.19.0-cp38-cp38-manylinux2010_x86_64.whl
Building wheels for collected packages: catwoman
  Building wheel for catwoman (PEP 517) ... done
  Created wheel for catwoman: filename=catwoman-1.0.7-cp38-cp38-linux_x86_64.whl size=111312 sha256=6e8483195977d501a8c37b043949d48d4be8d351a51100779abf2734163a3d88
  Stored in directory: /home/gb/.cache/pip/wheels/c4/b7/81/5eeddbc85db5c605f9f3182cd301ee94de1831a10294f5eaa4
Successfully built catwoman
Installing collected packages: numpy, catwoman
Successfully installed catwoman-1.0.7 numpy-1.19.0
$ python -c "import catwoman; print(catwoman.__version__)"
1.0.7
```
