### 2020-07-24 KathrynJones1/catwoman#4 @jessie-dotson
Tutorial creates error with suggested imports
The tutorial (https://catwoman.readthedocs.io/en/latest/tutorial.html) specifies
import matplotlib as plt

this causes an error in the third cell
<ipython-input-9-efc55d2707d4> in <module>
----> 1 plt.plot(time, flux)
      2 plt.xlabel("Time from central transit/days")
      3 plt.ylabel("Relative flux")
      4 plt.show()

AttributeError: module 'matplotlib' has no attribute 'plot'

There are several fixes, but changing to
import matplotlib.pyplot as plt
As done in the quickstart fixes it.
