### 2020-07-24 #2382 @jessie-dotson
Created https://github.com/KathrynJones1/catwoman/issues/4 to capture an easy to fix error in the tutorial.
