### 2020-07-24 KathrynJones1/catwoman#5 @jessie-dotson
typo in tutorial (mod vs model)
The tutorial initializes the lightcurve model into the variable "model", but subsequently demonstrates how to modify the model without reinitializing as
flux2 = mod.light_curve(params)         #calculates light curve

This creates an error because mod doesn't exist.  Replacing with model.light_curve(params) fixes the issue.
