### 2020-07-24 #2382 @jessie-dotson
Created https://github.com/KathrynJones1/catwoman/issues/5 to capture another easy to fix error in the tutorial.
