### 2020-07-24 KathrynJones1/catwoman#6 @jessie-dotson
missing () in tutorial
At the beginning of the example of how to loop over parameters (cool functionality btw!), the line
flux = np.zeros(7,len(time))

results in an error:
<ipython-input-23-902820660ea9> in <module>
----> 1 flux = np.zeros(7,len(time))
      2 params.rp = 0.1
      3 params.rp2 = 0.15
      4
      5 for i in range(0,7):

TypeError: data type not understood

Easily fixed by adding the necessary parens around the declaration of the shape fo the desired flux variable:
flux = np.zeros((7,len(time)))
