### 2020-07-24 KathrynJones1/catwoman#7 @jessie-dotson
tutorial calls undefined variable t in a plot command
In the cell where residuals are plotted, the following error occurs:
NameError                                 Traceback (most recent call last)
<ipython-input-27-7407140266de> in <module>
      1 for i in range(1,7):
----> 2         plt.plot(t,flux[i]-flux[0],label=str((i-3)*30)+'° - 0°')
      3         plt.xlabel("Time from central transit/days")
      4 plt.ylabel("Relative flux")
      5 plt.legend()

NameError: name 't' is not defined

the following command works:
plt.plot(time,flux[i]-flux[0],label=str((i-3)*30)+'° - 0°')
