### 2020-07-24 #2382 @jessie-dotson
Another easy to fix tutorial error is captured in https://github.com/KathrynJones1/catwoman/issues/7
