### 2020-07-24 KathrynJones1/catwoman#10 @jessie-dotson
List of dependencies not included in documentation
The pip install worked like a charm for me.  But JOSS recommends that the documentation include a list of dependencies.  Recommend adding that list to the install documentation page.  (With other packages it's been super-useful to have that list include the minimum version # of the required packages as well as just the list...)
