### 2020-07-24 KathrynJones1/catwoman#11 @jessie-dotson
Could not find automated tests or description of manual tests
JOSS asks:  "Are there automated tests or manual steps described so that the functionality of the software can be verified?"  I didn't find any automated tests or explicit manual tests.   Reproducing the tutorial enabled me to verify basic functionality, so that is a manual test of sorts.  However the tutorial didn't include example results to compare to for the supersampling functionality or the impact of adjusting the error tolerance.

This isn't a blocking issue given that there are effectively basic manual tests, but an automated or more explicit manual test suite would be beneficial to the package.
