### 2020-07-25 #2382 @KathrynJones1
@jessie-dotson Thank you for finding those typos, they should all be fixed now :)
