### 2020-07-27 #2382 @jessie-dotson
@KathrynJones1 awesome!  Nice job jumping right on those!  You might want to go ahead and close those issues in your catwoman github repository.  That way if anyone were following the trail from here to there, it'd be really really clear the typos were all fixed!

Also, I opened a couple other issues in your repository.  I didn't link them here because (imho) they aren't blockers for the JOSS review, but they are suggestions motivated by the JOSS checklist I'm working through.  If they're easy to address, please consider doing so.
