### 2020-07-27 #2382 @jessie-dotson
The text of the paper has a performance metric which I interpret as saying that it takes ~ 340 seconds to generate 1 million limb-darkened lightcurves each with 100 data points.  I can't reproduce that metric.   I have a different processor (2.6 GHz Intel Core i7) and am finding it takes about 4x longer.  The case I ran took 1415 seconds to create 1 million light curves with 100 data points.    Admittedly, I almost never benchmark code -- so it's *entirely* possible I didn't follow standard practices here.   The notebook showing my test is attached in this zip file.

[timing catwoman.ipynb.zip](https://github.com/openjournals/joss-reviews/files/4984187/timing.catwoman.ipynb.zip)
