### 2020-07-29 #2382 @KathrynJones1
@jessie-dotson I've had a look and I think firstly one of the reasons why your code took longer is that you set the `max_err=0.1`ppm however I calculated the timing based on the default setting of 1ppm instead (I could add this to the paper text to make it more explicit?) . It also seems you used a different timing method (time.process_time) whereas I used the timeit.default_timer() function (https://www.geeksforgeeks.org/timeit-python-examples/) which I tested actually gives a different (and lower) runtime for 1 mil lightcurves (i also found this which attempts to explain why there's a difference between the two: https://www.webucator.com/blog/2015/08/python-clocks-explained/). It seems that from what I've read the timeit.default_timer() may be better to use for timing sections of code although I admit I don't know much about this so happy to discuss in an new issue?
