### 2020-08-07 KathrynJones1/catwoman#11 @barentsen
I note that `catwoman` does ship with a `catwoman.tests.test()` function which fails when I run it:

```
$ python -c "from catwoman import tests; tests.test()"

Starting tests...

Testing uniform limb darkening model...
Traceback (most recent call last):
  File "<string>", line 1, in <module>
  File "/home/gb/tmp/tmpenv/lib/python3.8/site-packages/catwoman/tests.py", line 56, in test
    m = TransitModel(params, t)
  File "/home/gb/tmp/tmpenv/lib/python3.8/site-packages/catwoman/transitmodel.py", line 79, in __init__
    raise Exception("Incorrect number of coefficients for " +params.limb_dark + " limb darkening; u should have the form:\n \
Exception: Incorrect number of coefficients for uniform limb darkening; u should have the form:
 			 u = [] for uniform LD
 			 u = [u1] for linear LD
   			 u = [u1, u2] for quadratic, logarithmic, exponential, squareroot, and power2 LD
 			 u = [u1, u2, u3, u4] for nonlinear LD, or
 		         u = [u1, ..., un] for custom LD
```

Although automated tests are only *strongly encouraged* rather than *required* per JOSS guidelines, I do believe it is a blocking issue that the package contains a `tests.test()` function which fails.

**Finding:** I believe the `tests` module should either be removed, or it should be fixed and have its use included in the documentation.  I strongly encourage the authors to follow the latter approach, if possible!

(Tagging openjournals/joss-reviews#2382.)
