### 2020-08-07 #2382 @barentsen
In relation to the automated tests: I do believe it is a blocking issue that the package ships with a `catwoman.tests.test()` function which fails when I run it. Although automated tests are only strongly encouraged rather than required per JOSS guidelines, I do believe that if a package decides to ship with a `tests.test()` function then that function should succeed and be documented.

More details here: https://github.com/KathrynJones1/catwoman/issues/11#issuecomment-670287671
