### 2020-08-10 KathrynJones1/catwoman#11 @KathrynJones1
Hi there, thank you for pulling this up and for your recommendations. In the latest release (v 1.0.9) I have removed the redundant tests module so that it is hopefully no longer a blocking issue?
