### 2020-08-10 KathrynJones1/catwoman#10 @jessie-dotson
Ok, sorry to back on this one...  But I do think the package needs to list the minimum required version number for the dependencies.  I installed catwoman on another laptop that I only use occasionally.  It had an older version of numpy (1.16.2).  Catwoman installed cleanly, but it wouldn't run.   It threw an error when called catwoman.TransitModel.  Updating my numpy fixed the error.
