### 2020-08-10 #2382 @jessie-dotson
Please add minimum version#s of dependencies.  See discussion in https://github.com/KathrynJones1/catwoman/issues/10
