### 2020-08-10 #2382 @jessie-dotson
@KathrynJones1 -- thanks for the hint to use timeit!  I re-ran the performance test using timeit.  I still can't reproduce the quoted 300seconds, but get 400seconds instead.  I naively would've thought that my laptop (2.6 GHz Intel Core i7) would've been faster than the one quoted in the paper, but perhaps that's a naive expectation.  I also tried on an older processor (1.2 GHz Intel Core m5) and obtained ~600 seconds.  Since these are all within a factor of 2 and I don't have access to the specified processor, I'm calling it close enough...
