### 2020-08-12 KathrynJones1/catwoman#10 @KathrynJones1
Hi @jessie-dotson , so I've had a look and I've edited the pyproject.toml file (that was recommended in https://github.com/KathrynJones1/catwoman/issues/2 ) so that it specifies that the numpy version is >=1.16.2, which is the version I have on my laptop at the moment and it is able to run `catwoman` (so I'm not sure why the same version wouldn't work on yours?) . I've tested that if you try and download the latest `catwoman` (version 1.0.11) then it will reinstall an updated numpy. Do you think this is sufficient as an automatic process or do you think it requires an explicit list?
