### 2020-08-17 #2382 @barentsen
I completed my review and am happy for `catwoman` to be accepted by JOSS. :+1:

For future versions, I would encourage the authors to consider the following improvements:
* Add one or more automated unit tests.
* Expand the documentation to include more details on the following topics:
  * How to build the package from source.
  * How to contribute to the project.
  * Ideas for future improvements.
