### 2020-08-20 #2382 @jessie-dotson
I also have completed my review and am happy for catwoman to be accepted by JOSS.  The future improvements identified by @barentsen would be extremely useful!  Please do keep those in mind for the future.
