### 2020-08-20 KathrynJones1/catwoman#10 @jessie-dotson
Yes, please go ahead and close the issue!
