### 2020-09-11 #2382 @KathrynJones1
Okay so for v1.0.12, which I have labelled as the JOSS version the DOI from Zenodo is 10.5281/zenodo.3999513. Let me know if that is okay!
