### 2020-09-11 #2382 @arfon
@xuanxu @openjournals/joss-eics - just a reminder that we shouldn't publish this until we have the AAS (AJ) DOI.
