### 2020-09-25 #2382 @arfon
OK thanks @augustfly. In [this comment](https://github.com/openjournals/joss-reviews/issues/2330#issuecomment-643191751) @KathrynJones1 suggested an AJ submission was coming.

@KathrynJones1 - can you update us on your plans here?
