### 2020-09-25 #2382 @KathrynJones1
@arfon @augustfly yes we are in the final stages of preparing the other paper and plan to submit to AJ very soon. I will let you know when we do this
