### 2020-09-25 #2382 @xuanxu
@KathrynJones1 please merge [this pending pull request](https://github.com/KathrynJones1/catwoman/pull/14), it adds the missing journal names to the paper.
