### 2020-10-28 #2382 @arfon
:wave: @augustfly @KathrynJones1 - any updates on the status of the AJ paper here?
