### 2020-11-16 #2382 @KathrynJones1
Apologies for the delay, we have now submitted our AJ paper - do you need a DOI or are you happy with just a submission ID?
