### 2020-11-16 #2382 @augustfly
I see the manuscript, and I am marking it as a JOSS-related submission. We would not have an article DOI until the end of the peer review process (as Arfon knows I think).
