### 2020-11-16 #2382 @arfon
> I see the manuscript, and I am marking it as a JOSS-related submission. We would not have an article DOI until the end of the peer review process (as Arfon knows I think).

:+1: thanks. Yes, we just need the DOI from AAS publishing once the paper has been through review there.
