### 2020-11-16 #2382 @arfon
Actually, thinking this through, I think we should probably go ahead and publish this now and then update the JOSS metadata when the AAS paper is ready. What do you all think?
