### 2020-11-16 #2382 @augustfly
I'm find doing this in metadata.  This way @KathrynJones1 can update their manuscript at a later point to reference the JOSS article directly.
