### 2020-11-17 #2382 @arfon
@jessie-dotson, @barentsen - many thanks for your reviews here and to @xuanxu for editing this submission ✨. JOSS simply wouldn't work without the volunteer efforts of people like you.

Also, please let me know when the AAS paper has been accepted and you have a DOI, at that point we will associate this paper with the AAS one.
