### 2020-11-17 #2382 @whedon
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
