### 2021-08-19 #2382 @augustfly
@KathrynJones1 Did you end up submitting an AAS Journal article that couples to this JOSS publication? Is it https://arxiv.org/abs/2106.15687?
