### 2021-08-19 #2382 @augustfly
Great!  I'll catch up with the production office to get the DOI for the AJ paper. Our production office is running very behind, but I'm guessing based on the acceptance date that you and Nestor should have the proofs by now (or soon).
